requirements.txt
python-telegram-bot==20.7
google-generativeai
gtts
pygame
nest_asyncio
python-dotenv
